% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{add_epsilon}
\alias{add_epsilon}
\title{Helper to assist round_5_up}
\usage{
add_epsilon(x, epsilon = 1e-12)
}
\arguments{
\item{x}{[numeric] vector}

\item{epsilon}{[numeric] some small value to help e.g. 1225 round up to 1230}
}
\value{
[numeric] x +/- epsilon (for pos/neg x)
}
\description{
Helper to assist round_5_up
}
\seealso{
Other vector_formats: 
\code{\link{fmt_magnitude}()},
\code{\link{format_oxford_comma}()},
\code{\link{fround}()},
\code{\link{fround_count_rate}()},
\code{\link{fround_metric}()},
\code{\link{fround_metric_lancet}()},
\code{\link{fround_props}()}
}
\concept{vector_formats}
\keyword{internal}
