% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_data_type}
\alias{assert_data_type}
\title{Assert data type}
\usage{
assert_data_type(d_type)
}
\arguments{
\item{d_type}{[chr] data type to validate}
}
\value{
[chr] invisible validated d_type
}
\description{
Validates that a given data type is among the allowed types.
}
\seealso{
Other assertions: 
\code{\link{assert_clu_relationship}()},
\code{\link{assert_rate_unit}()},
\code{\link{assert_set_choice}()},
\code{\link{assert_style_schema}()},
\code{\link{assert_x_gte_y}()},
\code{\link{assert_x_in_y}()},
\code{\link{assert_x_not_in_y}()}
}
\concept{assertions}
\keyword{internal}
