% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_metric}
\alias{assert_metric}
\title{Assert metric}
\usage{
assert_metric(metric)
}
\arguments{
\item{metric}{[chr] metric type to validate}
}
\value{
[chr] invisible validated metric
}
\description{
Validates that a given metric is among the allowed types.
}
\seealso{
Other assertions: 
\code{\link{assert_clu_relationship}()},
\code{\link{assert_fround_return_schema}()},
\code{\link{assert_rate_unit}()},
\code{\link{assert_set_choice}()},
\code{\link{assert_style_schema}()},
\code{\link{assert_x_gte_y}()},
\code{\link{assert_x_in_y}()},
\code{\link{assert_x_not_in_y}()}
}
\concept{assertions}
\keyword{internal}
