% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_set_choice}
\alias{assert_set_choice}
\title{Assert set choice}
\usage{
assert_set_choice(x, choices)
}
\arguments{
\item{x}{[scalar] some scalar value}

\item{choices}{[vector] vector of allowed choices}
}
\value{
[scalar] invisible validated x
}
\description{
Assert set choice
}
\seealso{
Other assertions: 
\code{\link{assert_clu_relationship}()},
\code{\link{assert_data_type}()},
\code{\link{assert_style_schema}()},
\code{\link{assert_x_gte_y}()},
\code{\link{assert_x_in_y}()},
\code{\link{assert_x_not_in_y}()}
}
\concept{assertions}
