% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\encoding{UTF-8}
\name{assert_style_schema}
\alias{assert_style_schema}
\title{Assert style schema}
\usage{
assert_style_schema(style_entry)
}
\arguments{
\item{style_entry}{[list] named list representing a style entry}
}
\value{
[list] invisible validated style_entry
}
\description{
Validates that a style entry conforms to the expected schema defined in
get_style_schema().
}
\examples{
assert_style_schema(
   list(
      digits_round_prop          = 2
      , digits_sigfig_count      = 3
      , nsmall                   = 1
      , decimal.mark             = "."
      , negative_sign            = "-"
      , big.mark_count           = ","
      , mean_neg_text            = "a decrease of"
      , UI_only                  = FALSE
      , UI_text                  = ""
      , assert_clu_relationships = TRUE
      , is_lancet                = FALSE
   )
)
}
\seealso{
Other assertions: 
\code{\link{assert_clu_relationship}()},
\code{\link{assert_data_type}()},
\code{\link{assert_set_choice}()},
\code{\link{assert_x_gte_y}()},
\code{\link{assert_x_in_y}()},
\code{\link{assert_x_not_in_y}()}
}
\concept{assertions}
