% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\encoding{UTF-8}
\name{assert_x_gte_y}
\alias{assert_x_gte_y}
\title{Assert Greater Than or Equal To}
\usage{
assert_x_gte_y(x, y)
}
\arguments{
\item{x}{[num]eric vector}

\item{y}{[num]eric vector}
}
\value{
[none] stop if any elements of x are greater than y
}
\description{
Assert Greater Than or Equal To
}
\examples{
lower = c(1, 2, 3)
mean = c(3, 3, 3)
assert_x_gte_y(mean, lower)
mean = c(1, 2, 4)
# assert_x_gte_y(lower, mean) # stop
}
\seealso{
Other assertions: 
\code{\link{assert_clu_relationship}()},
\code{\link{assert_data_type}()},
\code{\link{assert_set_choice}()},
\code{\link{assert_style_schema}()},
\code{\link{assert_x_in_y}()},
\code{\link{assert_x_not_in_y}()}
}
\concept{assertions}
