% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_entry.R
\name{complete_entry}
\alias{complete_entry}
\title{Save and encrypt entry.}
\usage{
complete_entry(entry)
}
\arguments{
\item{entry}{journal entry object to be completed.}
}
\description{
Save and encrypt entry.
}
