% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_journal.R
\name{create_journal}
\alias{create_journal}
\title{Create a new journal.}
\usage{
create_journal(dir_path, password = "no")
}
\arguments{
\item{dir_path}{Directory of the new project.}

\item{password}{Character input. `"No"` if no password required,}
}
\description{
Create a new journal.
}
