% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_entry.Internal.R
\name{decrypt_entry.Internal}
\alias{decrypt_entry.Internal}
\title{Decrypt a log entry with the journal password.}
\usage{
decrypt_entry.Internal(file)
}
\arguments{
\item{file}{Path to encrypted log entry file.}
}
\description{
Decrypt a log entry with the journal password.
}
