% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mods.R
\name{drop_column}
\alias{drop_column}
\title{Drop a column from a data.frame or data.table, preserving class and
modifying in place if data.table.}
\usage{
drop_column(x, varname)
}
\arguments{
\item{x}{[data.frame or data.table]}

\item{varname}{[chr] column name to drop}
}
\value{
[data.frame or data.table] input `x` with column dropped
}
\description{
Drop a column from a data.frame or data.table, preserving class and
modifying in place if data.table.
}
\seealso{
Other column_mods: 
\code{\link{add_column}()},
\code{\link{drop_columns}()}
}
\concept{column_mods}
\keyword{internal}
