% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mods.R
\name{drop_columns}
\alias{drop_columns}
\title{Vectorized version of drop_column()}
\usage{
drop_columns(x, varnames)
}
\arguments{
\item{x}{[data.frame or data.table]}

\item{varnames}{[chr vector] column names to drop}
}
\value{
[data.frame or data.table] input `x` with columns dropped
}
\description{
Vectorized version of drop_column()
}
\examples{
df <- data.frame(a = 1:3, b = letters[1:3], c = c(TRUE, FALSE, TRUE))
df <- drop_columns(df, c("b", "c"))
}
\seealso{
Other column_mods: 
\code{\link{add_column}()},
\code{\link{drop_column}()}
}
\concept{column_mods}
