% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{en_dash}
\alias{en_dash}
\title{En-dash}
\usage{
en_dash()
}
\value{
[chr] en-dash character
}
\description{
Standard for "x -- y" and Lancet negative: ("\\U2013")
}
\examples{
en_dash()
}
\seealso{
Other marks: 
\code{\link{mid_dot}()},
\code{\link{thin_space}()}
}
\concept{marks}
