% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{flush_df_mag_state}
\alias{flush_df_mag_state}
\title{Flush df_mag state}
\usage{
flush_df_mag_state()
}
\value{
The df_mag data.frame that was stored (before clearing), or NULL
}
\description{
Flush df_mag state
}
\keyword{internal}
