% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{format_decimal}
\alias{format_decimal}
\title{Format value using decimal method}
\usage{
format_decimal(x_sc, nsmall, decimal.mark, big.mark)
}
\arguments{
\item{x_sc}{[num] scaled numeric value}

\item{nsmall}{[int] decimal places}

\item{decimal.mark}{[chr] decimal mark}

\item{big.mark}{[chr] thousands separator}
}
\value{
[chr] formatted string
}
\description{
non-exported helper
}
\keyword{internal}
