% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{format_int}
\alias{format_int}
\title{Format value using integer method}
\usage{
format_int(x_sc, decimal.mark, big.mark)
}
\arguments{
\item{x_sc}{[num] scaled numeric value}

\item{decimal.mark}{[chr] decimal mark}

\item{big.mark}{[chr] thousands separator}
}
\value{
[chr] formatted string
}
\description{
non-exported helper
}
\keyword{internal}
