% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journal.R
\name{format_lancet_df}
\alias{format_lancet_df}
\title{Return a table with formatted central, lower, upper for Lancet journal}
\usage{
format_lancet_df(
  df,
  d_type,
  new_var = "clu_fmt",
  central_var = "mean",
  lower_var = "lower",
  upper_var = "upper",
  remove_clu_columns = TRUE
)
}
\arguments{
\item{df}{[data.table] with central, lower, upper columns}

\item{d_type}{[chr c(prop', 'pp', 'count')] data type - proportion, percentage
point or count}

\item{new_var}{[chr: default 'clu_fmt'] name of new formatted column}

\item{central_var}{[chr: default 'mean'] name of central tendency e.g.
'point_estimate'}

\item{lower_var}{[chr: default 'lower']}

\item{upper_var}{[chr: default 'upper']}

\item{remove_clu_columns}{[lgl: default TRUE] remove central, lower, upper columns?}
}
\value{
[data.frame, data.table] with mean_95_UI_formatted column, and
  central, lower, upper columns removed (if specified)
}
\description{
Assumes a single data-type (d_type) for the whole table (e.g. 'prop', 'pp',
'count')
}
\examples{
df <- data.frame(
   location_did    = 1
   , location_name = "Global"
   , me_name       = "vacc_dpt1"
   , mean          = 55.8e6
   , lower         = 50.7e6
   , upper         = 60.7e6
)
format_lancet_df(df = df, d_type = "count", central_var = 'mean')
}
\seealso{
Other styled_formats: 
\code{\link{format_journal_clu}()},
\code{\link{format_journal_df}()},
\code{\link{format_lancet_clu}()},
\code{\link{format_means_df}()},
\code{\link{format_nature_clu}()},
\code{\link{format_nature_df}()},
\code{\link{fround_mag_clu}()}
}
\concept{styled_formats}
