% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journal.R
\name{format_nature_clu}
\alias{format_nature_clu}
\title{Format central, lower, upper value triplets for Nature journal presentation}
\usage{
format_nature_clu(central, lower, upper, d_type)
}
\arguments{
\item{central}{[num] central, point_estimate value vector}

\item{lower}{[num] lower bound vector}

\item{upper}{[num] upper bound vector}

\item{d_type}{[chr c(prop, pp, count)] data type - proportion, percentage}
}
\value{
[chr] formatted string vector
}
\description{
Format central, lower, upper value triplets for Nature journal presentation
}
\examples{
format_nature_clu(
   central  = c(0.994, -0.994)
   , lower  = c(0.984, -0.998)
   , upper  = c(0.998, -0.984)
   , d_type = "prop"
)
}
\seealso{
Other styled_formats: 
\code{\link{format_journal_clu}()},
\code{\link{format_journal_df}()},
\code{\link{format_lancet_clu}()},
\code{\link{format_lancet_df}()},
\code{\link{format_means_df}()},
\code{\link{format_nature_df}()},
\code{\link{fround_clu_triplet}()}
}
\concept{styled_formats}
