% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{format_oxford_comma}
\alias{format_oxford_comma}
\title{Format vector of items with Oxford comma}
\usage{
format_oxford_comma(vec, sep = "and")
}
\arguments{
\item{vec}{[any] vector of items to format}

\item{sep}{[chr: default "and"] separator before last item}
}
\value{
[chr] formatted string with Oxford comma
}
\description{
Format vector of items with Oxford comma
}
\examples{
format_oxford_comma(1:2)
format_oxford_comma(1:3)
format_oxford_comma(1:3, "or")
}
\seealso{
Other vector_formats: 
\code{\link{fmt_magnitude}()},
\code{\link{fround}()},
\code{\link{fround_count_rate}()},
\code{\link{fround_dtype}()},
\code{\link{fround_dtype_lancet}()},
\code{\link{fround_props}()}
}
\concept{vector_formats}
