% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{fround}
\alias{fround}
\title{Format and round}
\usage{
fround(x, digits = 1L, nsmall = 1L, decimal.mark = ".")
}
\arguments{
\item{x}{[num] numeric vector}

\item{digits}{[integer] passed to `round()`}

\item{nsmall}{[integer] passed to `format()`}

\item{decimal.mark}{[chr] passed to `format()`}
}
\value{
[chr] formatted string
}
\description{
Unaware of data-type or schema, just a hard-coded git-er-done function.
}
\examples{
fround(0.123456789)
fround(0.123456789, digits = 3)
fround(0.123456789, digits = 3, nsmall = 4)
}
\seealso{
Other vector_formats: 
\code{\link{fmt_magnitude}()},
\code{\link{format_oxford_comma}()},
\code{\link{fround_count_rate}()},
\code{\link{fround_dtype}()},
\code{\link{fround_dtype_lancet}()},
\code{\link{fround_props}()}
}
\concept{vector_formats}
