% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{fround_countish}
\alias{fround_countish}
\title{Format and round count-ish number}
\usage{
fround_countish(
  clu,
  mag_list,
  digits_sigfig_count = 3L,
  nsmall = 1L,
  decimal.mark = ".",
  big.mark_count = ",",
  is_lancet = FALSE
)
}
\arguments{
\item{clu}{[num] numeric triplet of counts (central, lower, upper)}

\item{mag_list}{[list] magnitude list as returned by `set_magnitude()`}

\item{digits_sigfig_count}{[integer] significant figures for counts}

\item{nsmall}{[integer] minimum number of digits to the right of the decimal point}

\item{decimal.mark}{[chr] passed to `format()`}

\item{big.mark_count}{[chr] passed to `format()`}

\item{is_lancet}{[lgl] if TRUE, apply Lancet-specific formatting rules}
}
\value{
[chr] formatted string vector
}
\description{
non-exported helper
}
\examples{
\dontrun{
fround_countish(
   clu                   = c(12345, 67890, 6e6)
   , mag_list            = set_magnitude(12345)
   , digits_sigfig_count = 3L
   , nsmall              = 1L
   , decimal.mark        = "."
   , big.mark_count      = ","
   , is_lancet           = FALSE
)
}
}
