% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_journal.R
\name{fround_mag_clu_nature}
\alias{fround_mag_clu_nature}
\title{Format and round central, lower, upper value sets by magnitude without units
for Nature journal presentation.}
\usage{
fround_mag_clu_nature(clu, d_type, df_mag = set_magnitude(clu[1]))
}
\arguments{
\item{clu}{[num] a numeric vector of three values in central, lower, upper order.}

\item{d_type}{[chr c('prop', 'pp', or 'count')] data type - proportion,
percentage point or count.}

\item{df_mag}{[data.frame] output from `set_magnitude()` - must be based on
central value of a central, lower, upper set - central and all UI values
inherit the same scale as the central tendency.}
}
\value{
[chr] formatted string (vectorized)
}
\description{
Format and round central, lower, upper value sets by magnitude without units
for Nature journal presentation.
}
\examples{
fround_mag_clu_nature(c(central = 0.2, lower = 0.1, upper = 0.3), "prop")
}
