% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{fround_metric}
\alias{fround_metric}
\title{Format and round with data-type suffix}
\usage{
fround_metric(x, metric = "prop", digits = 1L, nsmall = 1L, decimal.mark = ".")
}
\arguments{
\item{x}{[num] numeric value}

\item{metric}{[chr: default 'prop' c('prop', 'pp', 'count', 'rate')]}

\item{digits}{[integer: default 1L] passed to `round()`}

\item{nsmall}{[integer: default 1L] passed to `format()`}

\item{decimal.mark}{[chr: default "."] decimal mark passed to `format()`}
}
\value{
[chr] formatted string
}
\description{
Unaware of schema, just a hard-coded git-er-done function.
}
\examples{
fround_metric(0.123456789)
fround_metric(0.123456789, 'pp', 3, 4)
fround_metric(c(55.8346, 123.456789), 'count', 3, 4, ".")
}
\seealso{
Other vector_formats: 
\code{\link{add_epsilon}()},
\code{\link{fmt_magnitude}()},
\code{\link{format_oxford_comma}()},
\code{\link{fround}()},
\code{\link{fround_count_rate}()},
\code{\link{fround_metric_lancet}()},
\code{\link{fround_props}()}
}
\concept{vector_formats}
