% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{fround_propish}
\alias{fround_propish}
\title{Format and round proportion-ish number}
\usage{
fround_propish(x, digits = 1L, nsmall = 1L, decimal.mark = ".")
}
\arguments{
\item{x}{[num] numeric vector}

\item{digits}{[integer] passed to `round()`}

\item{nsmall}{[integer] passed to `format()`}

\item{decimal.mark}{[chr] passed to `format()`}
}
\value{
[chr] formatted string
}
\description{
non-exported helper
}
\examples{
\dontrun{
fround_propish(0.123456789)
}
}
