% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{fround_props}
\alias{fround_props}
\title{Format and round proportion-ish number}
\usage{
fround_props(clu, style_name)
}
\arguments{
\item{clu}{[num] numeric vector}

\item{style_name}{[chr] style name - controls rounding and
formatting.}
}
\value{
[chr] formatted string
}
\description{
non-exported helper
}
\details{
"well that was easy, how hard could counts be?"
}
\examples{
\dontrun{
fround_propish(0.123456789, 'nature')
}
}
\seealso{
Other vector_formats: 
\code{\link{fmt_magnitude}()},
\code{\link{format_oxford_comma}()},
\code{\link{fround}()},
\code{\link{fround_count}()},
\code{\link{fround_dtype}()},
\code{\link{fround_dtype_lancet}()}
}
\concept{vector_formats}
\keyword{internal}
