% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hex.R
\name{generate_hex}
\alias{generate_hex}
\title{Generates a string of randomly chosen characters.}
\usage{
generate_hex(length = 30, all_characters = TRUE)
}
\arguments{
\item{length}{Number of characters.}

\item{all_characters}{all chars or just alpha-numeric}
}
\description{
Generates a string of randomly chosen characters.
}
