% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{get_data_type_labels}
\alias{get_data_type_labels}
\title{Get data type labels}
\usage{
get_data_type_labels(d_type)
}
\arguments{
\item{d_type}{[chr]}
}
\value{
[list] named list of data type labels
}
\description{
Centrally managed definition for all data type labels.
}
\examples{
get_data_type_labels('prop')
}
\seealso{
Other data_types: 
\code{\link{get_data_types}()}
}
\concept{data_types}
