% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{get_data_types}
\alias{get_data_types}
\title{Get data types}
\usage{
get_data_types()
}
\value{
[chr] vector of allowed data types
}
\description{
Centrally managed definition for all allowed data types.
}
\examples{
get_data_types()
}
\seealso{
Other data_types: 
\code{\link{get_data_type_labels}()}
}
\concept{data_types}
