% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_df_mag_row}
\alias{get_df_mag_row}
\title{Get a single row from df_mag state}
\usage{
get_df_mag_row(idx)
}
\arguments{
\item{idx}{Row index (1-based)}
}
\value{
Single-row data.frame
}
\description{
Get a single row from df_mag state
}
\keyword{internal}
