% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_df_mag_state}
\alias{get_df_mag_state}
\title{Get current df_mag state}
\usage{
get_df_mag_state(env = get_dict_formats())
}
\arguments{
\item{env}{[env: default get_dict_formats()] Environment to get state from (defaults to package environment)}
}
\value{
df_mag [data.frame or NULL]
}
\description{
Get current df_mag state
}
\keyword{internal}
