% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_dict_format}
\alias{get_dict_format}
\title{Get dictionary by name}
\usage{
get_dict_format(dict_name)
}
\arguments{
\item{dict_name}{[chr] name of the dictionary to retrieve}
}
\value{
[any] the requested dictionary
}
\description{
Accessor function to retrieve a format dictionary from the package's
dictionary environment.
}
\examples{
\dontrun{
get_dict_format("lancet")
}
}
\keyword{internal}
