% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_dict_formats}
\alias{get_dict_formats}
\title{Get the package environment}
\usage{
get_dict_formats()
}
\value{
[env] the package's dictionary environment
}
\description{
For internal use.
}
\examples{
\dontrun{
get_dict_formats()
}
}
