% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_dict_formats_names}
\alias{get_dict_formats_names}
\title{Get all pre-assigned .dict_formats names}
\usage{
get_dict_formats_names()
}
\value{
[chr] names of all pre-assigned dictionaries in .dict_formats
}
\description{
Get all pre-assigned .dict_formats names
}
\keyword{internal}
