% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_metric_labels}
\alias{get_metric_labels}
\title{Get metric labels}
\usage{
get_metric_labels(metric)
}
\arguments{
\item{metric}{[chr]}
}
\value{
[list] named list of metric labels
}
\description{
Centrally managed definition for all metric labels.
}
\examples{
get_metric_labels('prop')
}
\seealso{
Other metrics: 
\code{\link{get_metrics}()}
}
\concept{metrics}
