% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get metrics}
\usage{
get_metrics()
}
\value{
[chr] vector of allowed metrics
}
\description{
Centrally managed definition for all allowed metrics.
}
\examples{
get_metrics()
}
\seealso{
Other metrics: 
\code{\link{get_metric_labels}()}
}
\concept{metrics}
