% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{get_style}
\alias{get_style}
\title{Get a style from the styles dictionary}
\usage{
get_style(style_name)
}
\arguments{
\item{style_name}{[chr] name of the style to retrieve}
}
\value{
[list] the requested style as a named list
}
\description{
Accessor function to retrieve a style from the package's
styles dictionary.
}
\examples{
get_style("lancet")
}
\seealso{
Other styles: 
\code{\link{assert_style_schema}()},
\code{\link{get_style_schema}()},
\code{\link{new_style}()},
\code{\link{set_style}()},
\code{\link{style_lancet}()},
\code{\link{style_nature}()}
}
\concept{styles}
