% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{get_style_item_by_data_type}
\alias{get_style_item_by_data_type}
\title{Helper for `format_means_df()`, perhaps others}
\usage{
get_style_item_by_data_type(style_name, style_item, d_type)
}
\arguments{
\item{style_name}{[chr] name of a style}

\item{style_item}{[chr] a style list item name}

\item{d_type}{[chr] a valid data type}
}
\value{
[scalar] some style item, type may vary
}
\description{
Helper for `format_means_df()`, perhaps others
}
