% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_style_item_by_metric}
\alias{get_style_item_by_metric}
\title{Helper for `format_means_df()`, perhaps others}
\usage{
get_style_item_by_metric(style_name, style_item, metric)
}
\arguments{
\item{style_name}{[chr] name of a style}

\item{style_item}{[chr] a style list item name}

\item{metric}{[chr] a valid metric type}
}
\value{
[scalar] some style item, type may vary
}
\description{
Helper for `format_means_df()`, perhaps others
}
\keyword{internal}
