% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{get_style_schema}
\alias{get_style_schema}
\title{Get style schema}
\usage{
get_style_schema()
}
\value{
[list] named list of style elements and their expected types
}
\description{
Centrally managed definition for all required journal format styles.
}
\examples{
get_style_schema()
}
\seealso{
Other styles: 
\code{\link{get_style}()},
\code{\link{new_style}()},
\code{\link{set_style}()},
\code{\link{style_lancet}()},
\code{\link{style_nature}()}
}
\concept{styles}
