% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{init_df_mag_state}
\alias{init_df_mag_state}
\title{Initialize df_mag state}
\usage{
init_df_mag_state(n)
}
\arguments{
\item{n}{Number of rows (triplets) to process}
}
\value{
NULL invisibly
}
\description{
Initialize df_mag state
}
\keyword{internal}
