% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is_df_mag_active}
\alias{is_df_mag_active}
\title{Check if df_mag state is active}
\usage{
is_df_mag_active(env = get_dict_formats())
}
\arguments{
\item{env}{[env: default get_dict_formats()] Environment to check state in (defaults to package environment)}
}
\value{
Logical
}
\description{
Check if df_mag state is active
}
\keyword{internal}
