% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{lock_some_bindings}
\alias{lock_some_bindings}
\title{Lock select bindings in the global environment}
\usage{
lock_some_bindings(objs, env = globalenv())
}
\arguments{
\item{objs}{[chr] a vector of object names to lock}

\item{env}{[environment] the environment to lock the bindings in}
}
\value{
[chr] invisible vector of input objects, to allow easier un-locking
}
\description{
Safely tries to lock the specified object bindings in the given environment.
}
\concept{bindings}
\keyword{internal}
