% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorise_password.R
\name{memorise_password}
\alias{memorise_password}
\title{Use this function to memorise your password.}
\usage{
memorise_password(password = NULL)
}
\arguments{
\item{password}{Enter the correct password. If NULL, then a dialog opens to enter the password.}
}
\description{
Use this function to memorise your password.
}
