% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{mid_dot}
\alias{mid_dot}
\title{Mid-dot}
\usage{
mid_dot()
}
\value{
[chr] mid-dot character
}
\description{
Lancet numeric decimal standard
}
\examples{
mid_dot()
}
\seealso{
Other marks: 
\code{\link{en_dash}()},
\code{\link{thin_space}()}
}
\concept{marks}
