% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{new_style}
\alias{new_style}
\title{Create a new style}
\usage{
new_style(
  style_name,
  digits_round_prop,
  digits_sigfig_count,
  nsmall,
  decimal.mark,
  negative_sign,
  big.mark_count,
  mean_neg_text,
  UI_only,
  UI_text,
  assert_clu_relationships,
  is_lancet,
  allow_thousands
)
}
\arguments{
\item{style_name}{[chr] name of the style to set}

\item{digits_round_prop}{[int] number of digits to round proportions to}

\item{digits_sigfig_count}{[int] number of significant figures for counts}

\item{nsmall}{[int] minimum number of digits to the right of the decimal point}

\item{decimal.mark}{[chr] decimal mark}

\item{negative_sign}{[chr] character to use for negative sign}

\item{big.mark_count}{[chr] character to use for counts thousand, million, billion separator}

\item{mean_neg_text}{[chr] text to use when describing negative mean differences}

\item{UI_only}{[lgl] whether to format for UI only}

\item{UI_text}{[chr] text to use for UI formatting}

\item{assert_clu_relationships}{[lgl] whether to assert CLU relationships (ensure lower < central < upper)}

\item{is_lancet}{[lgl] whether the style is for Lancet formatting - controls specific edge-case behaviors}

\item{allow_thousands}{[lgl] whether format counts as e.g. 10,000 as '10 thousand'}
}
\value{
[chr] invisible vector of input objects, to allow easier un-locking
}
\description{
Wrapper function to create and set a new style in one step.
}
\examples{
new_style(
  style_name    = "my_style"
  , digits_round_prop        = 1
  , digits_sigfig_count      = 1
  , nsmall                   = 1
  , decimal.mark             = "."
  , negative_sign            = "-"
  , big.mark_count           = ","
  , mean_neg_text            = "a decrease of"
  , UI_only                  = FALSE
  , UI_text                  = ""
  , assert_clu_relationships = TRUE
  , is_lancet                = FALSE
  , allow_thousands          = FALSE
)
}
\seealso{
Other styles: 
\code{\link{get_style}()},
\code{\link{get_style_schema}()},
\code{\link{set_style}()},
\code{\link{style_lancet}()},
\code{\link{style_nature}()}
}
\concept{styles}
