% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{process_clu_triplet_negatives}
\alias{process_clu_triplet_negatives}
\title{Prepare central, lower, upper value triplets for journal presentation}
\usage{
process_clu_triplet_negatives(triplets, assert_clu_order = TRUE)
}
\arguments{
\item{triplets}{[matrix]  with rownames 'central', 'lower', 'upper'}

\item{assert_clu_order}{[lgl: default TRUE] assert that central, lower, upper relationships are valid}
}
\value{
[num matrix] matrix with rows 'central', 'lower', 'upper' and columns for each triplet set
}
\description{
Vectorized preparation of central, lower, upper values.
Handles negatives, and swaps ordering where necessary.
Casting some negatives as positives allows user control to set
`style$neg_str_mean` appropriately.
}
\details{
non-exported helper
}
\examples{
\dontrun{
process_clu_triplet_negatives(
   central  = c(0.5, -0.2, -0.5, -2)
   , lower  = c(0.3, -0.5, -1.0, -3)
   , upper  = c(0.7, 0.1, -0.2, -1)
)
}
}
