% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{return_data_type_labels}
\alias{return_data_type_labels}
\title{Get data type labels}
\usage{
return_data_type_labels()
}
\value{
[list] named list of data type labels
}
\description{
Centrally managed definition for all data type labels.
}
\examples{
get_data_type_labels()
}
\seealso{
Other data_types: 
\code{\link{get_data_types}()}
}
\concept{data_types}
