% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_df_mag_state}
\alias{set_df_mag_state}
\title{Store a complete df_mag data.frame in state}
\usage{
set_df_mag_state(df_mag)
}
\arguments{
\item{df_mag}{data.frame as returned by set_magnitude()}
}
\value{
NULL invisibly
}
\description{
Store a complete df_mag data.frame in state
}
\keyword{internal}
