% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_dict_format}
\alias{set_dict_format}
\title{Set dictionary by name}
\usage{
set_dict_format(dict_name, dict_entry)
}
\arguments{
\item{dict_name}{[chr] name of the dictionary to set}

\item{dict_entry}{[any] value to assign to the dictionary}
}
\value{
[chr] invisible vector of input objects, to allow easier un-locking
}
\description{
Mutator function to set a format dictionary in the package's
dictionary environment.
}
\details{
Dictionary entries should only be settable with the accessor functions.
}
\keyword{internal}
