% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_entry.R
\name{start_entry}
\alias{start_entry}
\title{Create entry file.}
\usage{
start_entry(type = "md")
}
\arguments{
\item{type}{What type of entry. Options are currently md and rmd}
}
\description{
Create entry file.
}
