% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style_lancet}
\alias{style_lancet}
\title{Lancet style schema}
\usage{
style_lancet()
}
\value{
[list] named list representing the lancet style
}
\description{
Pre-defined style schema for Lancet journal formatting
}
\examples{
style_lancet()
}
\seealso{
Other styles: 
\code{\link{get_style}()},
\code{\link{get_style_schema}()},
\code{\link{new_style}()},
\code{\link{set_style}()},
\code{\link{style_nature}()}
}
\concept{styles}
