% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style_nature}
\alias{style_nature}
\title{Nature style schema}
\usage{
style_nature()
}
\value{
[list] named list representing the nature style
}
\description{
The default style for the package.
}
\details{
Pre-defined style schema for Nature journal formatting.
}
\examples{
style_nature()
}
\seealso{
Other styles: 
\code{\link{get_style}()},
\code{\link{get_style_schema}()},
\code{\link{new_style}()},
\code{\link{set_style}()},
\code{\link{style_lancet}()}
}
\concept{styles}
