% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{thin_space}
\alias{thin_space}
\title{Thin space}
\usage{
thin_space()
}
\value{
[chr] thin space character
}
\description{
Lancet thin space separator for counts 10,000 -- 999,9999 instead of comma
",": ("\\U2009")
}
\examples{
thin_space()
}
\seealso{
Other marks: 
\code{\link{en_dash}()},
\code{\link{mid_dot}()}
}
\concept{marks}
