% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mods.R
\name{add_column}
\alias{add_column}
\title{Add/overwrite a column in a data.frame or data.table, preserving class and
modifying in place if data.table.}
\usage{
add_column(x, varname, vec, overwrite = FALSE)
}
\arguments{
\item{x}{[data.frame or data.table]}

\item{varname}{[chr] column name for new column}

\item{vec}{[any] vector of values to add as new column}

\item{overwrite}{[lgl: default FALSE] overwrite existing column if TRUE}
}
\value{
[data.frame or data.table] input `x` with new column added
}
\description{
Add/overwrite a column in a data.frame or data.table, preserving class and
modifying in place if data.table.
}
\seealso{
Other column_mods: 
\code{\link{drop_column}()},
\code{\link{drop_columns}()}
}
\concept{column_mods}
\keyword{internal}
