% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_x_gte_y}
\alias{assert_x_gte_y}
\title{Assert Greater Than or Equal To}
\usage{
assert_x_gte_y(x, y)
}
\arguments{
\item{x}{[num]eric vector}

\item{y}{[num]eric vector}
}
\value{
[none] stop if any elements of x are greater than y
}
\description{
Assert Greater Than or Equal To
}
\examples{
lower = c(1, 2, 3)
mean = c(3, 3, 3)
assert_x_gte_y(mean, lower)
mean = c(1, 2, 4)
# assert_x_gte_y(lower, mean) # stop
}
