% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\encoding{UTF-8}
\name{assert_x_in_y}
\alias{assert_x_in_y}
\title{Assert all elements of x are in y}
\usage{
assert_x_in_y(x, y)
}
\arguments{
\item{x}{[vector] some vector}

\item{y}{[vector] some vector}
}
\value{
[none] stop if any elements of x are not in y
}
\description{
Assert all elements of x are in y
}
\examples{
assert_x_in_y(c("a", "b"), c("a", "b", "c"))
}
\seealso{
Other assertions: 
\code{\link{assert_clu_relationship}()},
\code{\link{assert_data_type}()},
\code{\link{assert_set_choice}()},
\code{\link{assert_style_schema}()},
\code{\link{assert_x_gte_y}()},
\code{\link{assert_x_not_in_y}()}
}
\concept{assertions}
