% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{fround_count}
\alias{fround_count}
\title{Format and round count-ish number}
\usage{
fround_count(clu, style_name, df_mag)
}
\arguments{
\item{clu}{[num] numeric triplet of counts (central, lower, upper)}

\item{style_name}{[chr] style name - controls rounding and
formatting.}

\item{df_mag}{[data.frame] magnitude df as returned by `set_magnitude()`}
}
\value{
[chr] formatted string vector
}
\description{
non-exported helper
}
\examples{
\dontrun{
fround_countish(clu = c(12345, 67890, 6.6666e6), df_mag = set_magnitude(12345), style_name = 'nature')
}
}
\seealso{
Other vector_formats: 
\code{\link{fmt_magnitude}()},
\code{\link{format_oxford_comma}()},
\code{\link{fround}()},
\code{\link{fround_dtype}()},
\code{\link{fround_dtype_lancet}()},
\code{\link{fround_props}()}
}
\concept{vector_formats}
