% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{fround_count}
\alias{fround_count}
\title{Format and round count-ish number}
\usage{
fround_count(clu, style_name, idx = NULL)
}
\arguments{
\item{clu}{[num] numeric triplet of counts (central, lower, upper)}

\item{style_name}{[chr] style name - controls rounding and
formatting.}

\item{idx}{[int] row index when called from format_journal_clu(), NULL for standalone}
}
\value{
[chr] formatted string vector
}
\description{
non-exported helper
}
\examples{
\dontrun{
fround_count(clu = c(12345, 67890, 6.6666e6), style_name = 'nature', idx = 1)
}
}
\seealso{
Other vector_formats: 
\code{\link{fmt_magnitude}()},
\code{\link{format_oxford_comma}()},
\code{\link{fround}()},
\code{\link{fround_dtype}()},
\code{\link{fround_dtype_lancet}()},
\code{\link{fround_props}()}
}
\concept{vector_formats}
\keyword{internal}
