% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{fround_props}
\alias{fround_props}
\title{Format and round proportion-ish number}
\usage{
fround_props(clu, style_name, mag = NULL)
}
\arguments{
\item{clu}{[num] numeric triplet of proportions (central, lower, upper)}

\item{style_name}{[chr] style name - controls rounding and formatting}

\item{mag}{[chr: default NULL] magnitude override - see set_magnitude()
- For props/pp: "as-is" (no scaling, use values as provided)
- For counts: "t" (thousand), "m" (million), "b" (billion)
- For rates: "per10", "per100", "per1k", ..., "per10b"}
}
\value{
[list] with elements:
  - formatted: chr[3] - formatted central, lower, upper values
  - df_mag_row: data.frame[1,] - magnitude info (mag, mag_label, denom)
}
\description{
non-exported helper
}
\details{
"well that was easy, how hard could counts be?"
}
\seealso{
Other vector_formats: 
\code{\link{add_epsilon}()},
\code{\link{fmt_magnitude}()},
\code{\link{format_oxford_comma}()},
\code{\link{fround}()},
\code{\link{fround_count_rate}()},
\code{\link{fround_metric}()},
\code{\link{fround_metric_lancet}()}
}
\concept{vector_formats}
\keyword{internal}
