% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vectors.R
\name{fround_props}
\alias{fround_props}
\title{Format and round proportion-ish number}
\usage{
fround_props(clu, style_name)
}
\arguments{
\item{clu}{[num] numeric triplet of proportions (central, lower, upper)}

\item{style_name}{[chr] style name - controls rounding and formatting}
}
\value{
[list] with elements:
  - formatted: chr[3] - formatted central, lower, upper values
  - df_mag_row: data.frame[1,] - magnitude info (mag, mag_label, denom)
}
\description{
non-exported helper
}
\details{
"well that was easy, how hard could counts be?"
}
\examples{
\dontrun{
result <- fround_props(c(0.123, 0.100, 0.150), 'nature')
result$formatted  # chr[3]
result$df_mag_row # df[1,]
}
}
\seealso{
Other vector_formats: 
\code{\link{add_epsilon}()},
\code{\link{fmt_magnitude}()},
\code{\link{format_oxford_comma}()},
\code{\link{fround}()},
\code{\link{fround_count_rate}()},
\code{\link{fround_metric}()},
\code{\link{fround_metric_lancet}()}
}
\concept{vector_formats}
\keyword{internal}
