% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_df_mag_row}
\alias{get_df_mag_row}
\title{Get a single row from df_mag state}
\usage{
get_df_mag_row(idx, env = get_dict_formats())
}
\arguments{
\item{idx}{[int] Row index (1-based)}

\item{env}{[env: default get_dict_formats()] Environment to get state from (defaults to package environment)}
}
\value{
[data.frame] Single-row
}
\description{
Get a single row from df_mag state
}
\keyword{internal}
