% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{get_metric_labels}
\alias{get_metric_labels}
\title{Get data type labels}
\usage{
get_metric_labels(metric)
}
\arguments{
\item{metric}{[chr]}
}
\value{
[list] named list of data type labels
}
\description{
Centrally managed definition for all data type labels.
}
\examples{
get_metric_labels('prop')
}
\seealso{
Other metrics: 
\code{\link{get_metrics}()}
}
\concept{metrics}
