% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{get_metrics}
\alias{get_metrics}
\title{Get data types}
\usage{
get_metrics()
}
\value{
[chr] vector of allowed data types
}
\description{
Centrally managed definition for all allowed data types.
}
\examples{
get_metrics()
}
\seealso{
Other metrics: 
\code{\link{get_metric_labels}()}
}
\concept{metrics}
