% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{init_df_mag_state}
\alias{init_df_mag_state}
\title{Initialize df_mag state}
\usage{
init_df_mag_state(n, env = get_dict_formats())
}
\arguments{
\item{n}{[int] Number of rows (triplets) to process}

\item{env}{[env: default get_dict_formats()] Environment to store state in (defaults to package environment)}
}
\value{
NULL invisibly
}
\description{
Initialize df_mag state
}
\keyword{internal}
