% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{new_style}
\alias{new_style}
\title{Make a new style by args}
\usage{
new_style(
  style_name,
  digits_round_prop,
  nsmall_prop,
  method_count,
  digits_sigfig_count,
  pad_count_sigfigs,
  nsmall_count,
  decimal.mark,
  neg_str_UI,
  big.mark_count,
  neg_str_mean,
  UI_only,
  UI_text,
  assert_clu_order,
  is_lancet,
  label_thousands,
  round5up
)
}
\arguments{
\item{style_name}{[chr] name of the style to set}

\item{digits_round_prop}{[int] number of digits to round proportions to}

\item{nsmall_prop}{[int] minimum number of digits to the right of the decimal point - proportions}

\item{digits_sigfig_count}{[int] number of significant figures for counts}

\item{decimal.mark}{[chr] decimal mark}

\item{neg_str_UI}{[chr] character to use for negative sign}

\item{big.mark_count}{[chr] character to use for counts thousand, million, billion separator}

\item{neg_str_mean}{[chr] text to use when describing negative mean differences}

\item{UI_only}{[lgl] whether to format for UI only}

\item{UI_text}{[chr] text to use for UI formatting}

\item{assert_clu_order}{[lgl] whether to assert CLU relationships (ensure lower < central < upper)}

\item{is_lancet}{[lgl] whether the style is for Lancet formatting - controls specific edge-case behaviors}

\item{label_thousands}{[lgl] whether format counts as e.g. 10,000 as '10 thousand'}
}
\value{
[chr] invisible vector of input objects, to allow easier un-locking
}
\description{
Wrapper function to create and set a new style in one step.
}
\examples{
new_style(
  style_name            = "my_style"
  , digits_round_prop   = 1
  , digits_sigfig_count = 1
  , pad_count_sigfigs   = TRUE
  , nsmall_prop         = 1
  , nsmall_count        = 1
  , decimal.mark        = "."
  , neg_str_UI          = "-"
  , big.mark_count      = ","
  , neg_str_mean        = "a decrease of"
  , UI_only             = FALSE
  , UI_text             = ""
  , assert_clu_order    = TRUE
  , is_lancet           = FALSE
  , label_thousands     = FALSE
  , round5up            = TRUE
)
}
\seealso{
Other styles: 
\code{\link{get_style}()},
\code{\link{get_style_schema}()},
\code{\link{set_style}()},
\code{\link{style_lancet}()},
\code{\link{style_nature}()}
}
\concept{styles}
