% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{set_style}
\alias{set_style}
\title{Set a new style schema}
\usage{
set_style(style_name, style_entry)
}
\arguments{
\item{style_name}{[chr] name of the style to set}

\item{style_entry}{[list] named list representing the style entry}
}
\value{
[chr] invisible vector of input objects, to allow easier un-locking
}
\description{
Set a new style schema
}
\examples{
set_style(
   style_name    = "my_style"
   , style_entry = list(
      digits_round_prop          = 2
      , digits_sigfig_count      = 3
      , nsmall                   = 1
      , decimal.mark             = "."
      , neg_str_UI            = "-"
      , big.mark_count           = ","
      , neg_str_mean            = "a decrease of"
      , UI_only                  = FALSE
      , UI_text                  = ""
      , assert_clu_order = TRUE
      , is_lancet                = FALSE
      , label_thousands          = FALSE
   )
)
}
\seealso{
Other styles: 
\code{\link{get_style}()},
\code{\link{get_style_schema}()},
\code{\link{new_style}()},
\code{\link{style_lancet}()},
\code{\link{style_nature}()}
}
\concept{styles}
